requirements.txt
torch>=1.9.0
torchvision>=0.10.0
rasterio>=1.2.0
numpy>=1.21.0
matplotlib>=3.4.0
opencv-python>=4.5.0
tqdm>=4.62.0
piq>=0.7.0
torchmetrics>=0.11.0
pillow>=8.3.0
scikit-image>=0.19.0
